+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/SPORatifySpec.hs
@@ -10,7 +10,6 @@
 
 module Test.Cardano.Ledger.Conway.SPORatifySpec (spec) where
 
-import Cardano.Ledger.Address (RewardAccount (..))
 import Cardano.Ledger.BaseTypes (StrictMaybe (..))
 import Cardano.Ledger.Coin (Coin (..), CompactForm (..))
 import Cardano.Ledger.Compactible (Compactible (..))
@@ -275,7 +274,7 @@ genTestData Ratios {yes, no, abstain, alwaysAbstain, noConfidence} = do
       Map (KeyHash StakePool) StakePoolState ->
       Map (Credential Staking) DRep
     mkDelegatees drep =
-      fromKeys (const drep) . map (raCredential . spsRewardAccount) . Map.elems
+      fromKeys (const drep) . map spsRewardAccount . Map.elems
 
     -- Create a map from each pool with the given value, where the key is the pool credential
     -- and take the union of all these maps.