+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Snap.hs
@@ -16,7 +16,7 @@ module Cardano.Ledger.Shelley.Rules.Snap (
   SnapEnv (..),
 ) where
 
-import Cardano.Ledger.BaseTypes (ShelleyBase)
+import Cardano.Ledger.BaseTypes (ShelleyBase, networkId)
 import Cardano.Ledger.Coin (Coin, CompactForm)
 import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Core
@@ -28,11 +28,13 @@ import Cardano.Ledger.Shelley.LedgerState (
  )
 import Cardano.Ledger.State
 import Control.DeepSeq (NFData)
+import Control.Monad.Trans.Reader (asks)
 import Control.State.Transition (
   STS (..),
   TRC (..),
   TransitionRule,
   judgmentContext,
+  liftSTS,
   tellEvent,
  )
 import Data.Map.Strict (Map)
@@ -73,14 +75,18 @@ instance (EraTxOut era, EraStake era, EraCertState era) => STS (ShelleySNAP era)
 --
 -- but is now computed incrementally. We leave the comment as a historical note about
 -- where important changes were made to the source code.
-snapTransition :: (EraStake era, EraCertState era) => TransitionRule (ShelleySNAP era)
+snapTransition ::
+  (EraStake era, EraCertState era, STS (ShelleySNAP era)) => TransitionRule (ShelleySNAP era)
 snapTransition = do
   TRC (snapEnv, s, _) <- judgmentContext
+
+  network <- liftSTS $ asks networkId
 
   let SnapEnv ls@(LedgerState (UTxOState _utxo _ fees _ _ _) certState) _pp = snapEnv
       instantStake = ls ^. instantStakeG
       -- per the spec: stakeSnap = stakeDistr @era utxo dstate pstate
-      istakeSnap = snapShotFromInstantStake instantStake (certState ^. certDStateL) (certState ^. certPStateL)
+      istakeSnap =
+        snapShotFromInstantStake instantStake (certState ^. certDStateL) (certState ^. certPStateL) network
 
   tellEvent $
     let stMap :: Map (Credential Staking) (CompactForm Coin)