+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/UnitTests/InstantStakeTest.hs
@@ -86,7 +86,8 @@ instantStakeIncludesRewards = do
     instantStake = addInstantStake utxo1 mempty
     poolparamMap = Map.fromList [(poolId1, pool1), (poolId2, pool2)]
   pState <- arbitraryLens psStakePoolsL $ mkStakePoolState mempty mempty <$> poolparamMap
-  let snapShot = snapShotFromInstantStake instantStake dState pState
+  network <- arbitrary
+  let snapShot = snapShotFromInstantStake instantStake dState pState network
       computedStakeDistr = VMap.toMap (unStake (ssStake snapShot))
 
       expectedStakeDistr :: Map (Credential Staking) (CompactForm Coin)